backend/requirements.txt
fastapi
uvicorn
python-multipart
# google-generativeai # Not used anymore
python-dotenv
sentence-transformers
numpy
openai
scikit-learn
